connect earlier, be more tolerant with Modbus devices
This tries to improve working with awkward modbus devices that react
slowly or don't always answer in time.
Motivation for these changes are Huawei SUN inverters. For SmartHome
purposes they can only be accessed via Modbus. However, its Modbus
implementation seems to be limited and rather restrictive because, e.g.,
only one client at a time can communicate with the inverter's Modbus
interface and it needs some time upon connection to be able to process
requests. This conflicts with modbus-proxy-rs's approach to only
connect when the first request is sent to the Modbus device.

This introduces to changes to fix the issues:
- directly connect to Modbus device upon first client connection.
  Clients that work with the Huawei SUN Modbus usually use some kind of
  time delay between connecting and sending their first request. So
  don't break that
- retry on connection or request failure. If connecting to the Modbus
  device fails or no answer has been received for a request, just try
  again up to 5 times. This handles the unreliable behaviour of some
  Modbus devices much more tolerantly.

Signed-off-by: Jonas Jelonek <jonas.jelonek@proton.me>